requirements.txt
aioboto3
boto3
python-dotenv
mcp[cli]